+++ src/contexts/privateWalletContext.tsx
@@ -184,7 +184,7 @@ export const PrivateWalletContextProvider = (props) => {
         setTxStatus(TxStatus.processing(null, lastTx.hash.toString()));
       } catch (e) {
         console.error('Error publishing private transaction batch', e);
-        setTxStatus(TxStatus.failed());
+        setTxStatus(TxStatus.failed('Transaction declined'));
         removePendingHistoryEvent();
         txQueue.current = [];
       }
@@ -232,11 +232,11 @@ export const PrivateWalletContextProvider = (props) => {
       externalAccount.address
     );
     if (signResult === null) {
-      return false;
+      setTxStatus(TxStatus.failed('Transaction declined'));
+      return;
     }
     const batches = signResult.txs;
-    const res = await publishBatchesSequentially(batches, txResHandler);
-    return res;
+    await publishBatchesSequentially(batches, txResHandler);
   };
 
   const privateTransfer = async (balance, recipient, txResHandler) => {
@@ -248,11 +248,11 @@ export const PrivateWalletContextProvider = (props) => {
       externalAccount.address
     );
     if (signResult === null) {
-      return false;
+      setTxStatus(TxStatus.failed('Transaction declined'));
+      return;
     }
     const batches = signResult.txs;
-    const res = await publishBatchesSequentially(batches, txResHandler);
-    return res;
+    await publishBatchesSequentially(batches, txResHandler);
   };
 
   const toPrivate = async (balance, txResHandler) => {
@@ -263,11 +263,11 @@ export const PrivateWalletContextProvider = (props) => {
       externalAccount.address
     );
     if (signResult === null) {
-      return false;
+      setTxStatus(TxStatus.failed('Transaction declined'));
+      return;
     }
     const batches = signResult.txs;
-    const res = await publishBatchesSequentially(batches, txResHandler);
-    return res;
+    await publishBatchesSequentially(batches, txResHandler);
   };
 
   const value = {