+++ src/types/TxStatus.ts
@@ -23,8 +23,8 @@ export default class TxStatus {
     return new TxStatus(FINALIZED, extrinsic, subscanUrl);
   }
 
-  static failed() {
-    return new TxStatus(FAILED);
+  static failed(message) {
+    return new TxStatus(FAILED, null, null, message);
   }
 
   isProcessing() {